% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_scores_values.R
\docType{data}
\name{alignment_scores_values}
\alias{alignment_scores_values}
\title{Help dataset for plotting alignment scores on an axis}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 5 columns.
}
\usage{
alignment_scores_values
}
\description{
Dataset containing numeric values corresponding to alphabetic scores for
easier plotting.
}
\examples{
pacta.executive.summary:::alignment_scores_values
}
\keyword{internal}
