% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_pdf.R
\name{blank_pdf}
\alias{blank_pdf}
\title{Get the path to the blank template PDF}
\usage{
blank_pdf()
}
\value{
A single string containing the path to a blank template PDF.
}
\description{
Get the path to the blank template PDF
}
