% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{lookup}
\alias{lookup}
\alias{time_horizon_lookup}
\title{Lookup valid values}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
time_horizon_lookup
}
\description{
Lookup valid values
}
\keyword{datasets}
