% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p4i_p4b_sector_technology_mapper}
\alias{p4i_p4b_sector_technology_mapper}
\title{Sector and technology names mapped from P4I to P4B style}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 4 columns.
}
\usage{
p4i_p4b_sector_technology_mapper
}
\description{
A data set containing the names of sectors and technologies according to
P4I and P4B conventions.
}
\keyword{datasets}
