% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.executive.summary-package.R
\docType{package}
\name{pacta.executive.summary-package}
\alias{pacta.executive.summary}
\alias{pacta.executive.summary-package}
\title{pacta.executive.summary: Creates Plots for the PACTA COP Executive Summary}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains plotting functions and a template for generating the executive summary that displays aggregated results of a PACTA COP exercise.
}
\author{
\strong{Maintainer}: Monika Furdyna \email{monika.furdyna@gmail.com} (\href{https://orcid.org/0000-0002-3728-0646}{ORCID}) [contractor]

}
\keyword{internal}
