% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignment_table.R
\name{plot_alignment_table}
\alias{plot_alignment_table}
\title{Create a scenario alignment table}
\usage{
plot_alignment_table(data)
}
\arguments{
\item{data}{A data frame. In principle, an output of
\code{prep_alignment_table()}. Requirements:
\itemize{
\item \code{asset_class} must have a single value.
\item Must have columns: \code{asset_class},\code{sector},\code{technology}, \code{entity},
\code{aligned_scen_temp}, \code{perc_aum}.
\item \code{entity} must contain at least one value of "portfolio".
\item \code{sector} must be one of: "power", "fossil_fuels", "automotive".
\item \code{aligned_scen_temp} must be one of: ">3.2C", "2.7-3.2C", "<2C".
\item \code{perc_aum} must be a percentage in decimal format, with values
between 0 and 1.
}}
}
\value{
an object of class "ggplot".
}
\description{
Create a scenario alignment table
}
\examples{
library(dplyr)

plot_alignment_table(toy_data_alignment_table \%>\% filter(asset_class == "equity"))
}
