% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagram.R
\name{plot_diagram}
\alias{plot_diagram}
\title{Create a diagram of asset class coverage}
\usage{
plot_diagram(data = NULL)
}
\arguments{
\item{data}{A data frame. In principle an output of \code{prep_diagram()}.
Requirements:
\itemize{
\item Must have columns: \code{asset_class}, \code{exposure_portfolio},
\code{exposure_asset_class}, \code{exposure_asset_class_perc}, \code{exposure_pacta},
\code{exposure_pacta_perc_asset_class_exposure}, \code{emissions_pacta_perc},
\code{emissions_pacta}.
\item Must have two rows.
\item \code{asset_class} must have values: "equity", "bonds".
\item \code{exposure_portfolio}, \code{exposure_asset_class}, \code{exposure_asset_class_perc},
\code{exposure_pacta}, \code{exposure_pacta_perc_asset_class_exposure},
\code{emissions_pacta_perc}, \code{emissions_pacta} must be numeric.
\item \code{exposure_asset_class_perc}, \code{exposure_pacta_perc_asset_class_exposure},
\code{emissions_pacta_perc} must be a percentage in decimal format, with values
between 0 and 1.
}}
}
\value{
An object of class "htmlwidget".
}
\description{
Create a diagram of asset class coverage
}
\examples{
plot_diagram(toy_data_diagram)
}
