% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emissions_scorecard.R
\name{plot_emissions_scorecard}
\alias{plot_emissions_scorecard}
\title{Create a plot showing scorecard emissions}
\usage{
plot_emissions_scorecard(data)
}
\arguments{
\item{data}{A data frame. In principle, an output of
\code{prep_emissions_scorecard()}. Requirements:
\itemize{
\item Must have columns: \code{asset_class}, \code{entity}, \code{emissions}.
\item \code{asset_class} must be one of: "equity", "bonds".
\item \code{entity} must be one of: "portfolio", "benchmark".
\item \code{emissions} must be numeric.
}}
}
\value{
an object of class "ggplot".
}
\description{
Create a plot showing scorecard emissions
}
\examples{
plot_emissions_scorecard(toy_data_emissions_scorecard)
}
