% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposures_scorecard.R
\name{plot_exposures_scorecard}
\alias{plot_exposures_scorecard}
\title{Create a plot showing exposure to sectors relevant to scorecard}
\usage{
plot_exposures_scorecard(data)
}
\arguments{
\item{data}{A data frame. In principle, an output of
\code{prep_exposures_scorecard()}. Requirements:
\itemize{
\item Must have columns: \code{asset_class}, \code{sector_or_tech}, \code{exposure_perc_aum}.
\item \code{sector_or_tech} must be one of: "coal", "other_fossil_fuels",
"fossil_power", "renewables_power".
\item \code{exposure_perc_aum} must be a percentage in decimal format, with values
between 0 and 1.
}}
}
\value{
an object of class "ggplot".
}
\description{
Create a plot showing exposure to sectors relevant to scorecard
}
\examples{
plot_exposures_scorecard(toy_data_exposures_scorecard)
}
