% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fossil_bars.R
\name{plot_fossil_bars}
\alias{plot_fossil_bars}
\title{Create a bar plot with exposures to fossil fuels}
\usage{
plot_fossil_bars(data)
}
\arguments{
\item{data}{A data frame. In principle, an output of \code{prep_fossil_bars()}.
Requirements:
\itemize{
\item Must have columns: \code{entity_name}, \code{entity_type},\code{asset_class},\code{tech},
\code{perc_aum}.
\item \code{tech} column must only have following values: "coal", "oil", "gas"
\item \code{perc_aum} must be a percentage in decimal format, with values between 0
and 1.
}}
}
\value{
an object of class "ggplot".
}
\description{
Create a bar plot with exposures to fossil fuels
}
\examples{
plot_fossil_bars(toy_data_fossil_bars)
}
