% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Create a scatter plot of exposure to low-carbon technology vs. alignment score}
\usage{
plot_scatter(data)
}
\arguments{
\item{data}{A data frame. In principle, an output of
\code{prep_scatter()}. Requirements:
\itemize{
\item \code{asset_class} must have a single value.
\item Must have columns: \code{asset_class},\code{tech_mix_green},\code{score}, \code{entity_name},
\code{entity_type}.
\item \code{entity_type} column must be factor and only have following values:
"average", "this_portfolio", "peer", "benchmark".
\item \code{tech_mix_green} must be a percentage in decimal format, with values
between 0 and 1.
\item \code{score} must be a number between 0 and 100.
}}
}
\value{
an object of class "ggplot".
}
\description{
Create a scatter plot of exposure to low-carbon technology vs. alignment score
}
\examples{
library(dplyr)

plot_scatter(toy_data_scatter \%>\% filter(asset_class == "equity"))
}
