% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{plot_scores}
\alias{plot_scores}
\title{Create a plot showing aggregated scores for portfolio and peers}
\usage{
plot_scores(data)
}
\arguments{
\item{data}{A data frame. In principle, an output of \code{prep_scores()}.
Requirements:
\itemize{
\item Must have columns: \code{asset_class},\code{scope},\code{entity}, \code{sector}, \code{score}.
\item \code{asset_class} must have a single value.
\item \code{scope} must be one of: "portfolio", "sector".
\item \code{entity} must have following values: "this_portfolio", "peers".
\item \code{sector} must have following values: "power", "automotive", "coal",
"oil", "gas", "steel", "cement" or \code{NA} in case of \code{scope} == "portfolio".
\item \code{score} must be one of: "A+", "A", "B", "C", "D", "E".
}}
}
\value{
an object of class "ggplot".
}
\description{
Create a plot showing aggregated scores for portfolio and peers
}
\examples{
library(dplyr)

plot_scores(toy_data_scores \%>\% filter(asset_class == "equity"))
}
