% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_scorecard.R
\name{plot_scores_scorecard}
\alias{plot_scores_scorecard}
\title{Create a plot showing aggregated score in scorecard}
\usage{
plot_scores_scorecard(data)
}
\arguments{
\item{data}{A data frame. In principle, an output of
\code{prep_scores_scorecard()}. Requirements:
\itemize{
\item Must have columns: \code{asset_class}, \code{score}.
\item \code{asset_class} must be one of: "equity", "bonds.
\item \code{score} must be one of: "A+", "A", "B", "C", "D", "E".
}}
}
\value{
an object of class "ggplot".
}
\description{
Create a plot showing aggregated score in scorecard
}
\examples{
data <- toy_data_scores \%>\%
dplyr::filter(
   scope == "portfolio",
   entity == "this_portfolio"
 )
plot_scores_scorecard(data)
}
