% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_alignment_table.R
\name{prep_alignment_table}
\alias{prep_alignment_table}
\title{Prepare data input for plotting technology alignment table}
\usage{
prep_alignment_table(
  results_portfolio = NULL,
  peers_results_aggregated = NULL,
  scenario_source = "GECO2021"
)
}
\arguments{
\item{results_portfolio}{Data frame that contains pre-wrangled portfolio
level PACTA results from a PACTA for investors analysis.}

\item{peers_results_aggregated}{Data frame that contains pre-wrangled
aggregate peer group level PACTA results from a PACTA for investors
analysis.}

\item{scenario_source}{Character. Must be a
\code{scenario_source} featured in the \code{scenario_thresholds} data set.}
}
\value{
data.frame
}
\description{
Prepare data input for plotting technology alignment table based
on PACTA for investors output files. These files must have been wrangled with
\code{prep_data_executive_summary()} before they can be passed to this function.
}
