% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_climate_strategy_scorecard.R
\name{prep_climate_strategy_scorecard}
\alias{prep_climate_strategy_scorecard}
\title{Prepare data input for climate strategy metrics in the scorecard section}
\usage{
prep_climate_strategy_scorecard(
  user_climate_strategy_scorecard,
  peers_climate_strategy_scorecard,
  peer_group = c("pensionfund", "assetmanager", "bank", "insurance", "other")
)
}
\arguments{
\item{user_climate_strategy_scorecard}{Data frame that contains
pre-calculated climate strategy metrics on the user level, based on the
responses to the COP survey.}

\item{peers_climate_strategy_scorecard}{Data frame that contains
pre-calculated climate strategy metrics on the peer group level, based on
the responses to the COP survey.}

\item{peer_group}{Character. Peer group of the analysed portfolio.}
}
\value{
data.frame
}
\description{
Prepare data input for climate strategy metrics in the scorecard section
based on COP survey results. The input data sets are pre-calculated and
accessed via the directories pertaining to the given user_id.
}
