% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_climate_strategy_scorecard.R
\name{prep_climate_strategy_scorecard_initiatives}
\alias{prep_climate_strategy_scorecard_initiatives}
\title{Prepare data input for climate strategy metrics (initiatives) in the scorecard}
\usage{
prep_climate_strategy_scorecard_initiatives(
  data,
  data_peers,
  peer_group = c("pensionfund", "assetmanager", "bank", "insurance", "other")
)
}
\arguments{
\item{data}{Data frame that contains pre-calculated climate strategy info on
participation in initiatives on the individual user level, based on the
responses to the COP survey.}

\item{data_peers}{Data frame that contains pre-calculated climate strategy
info on peer level participation in initiatives, based on the responses to
the COP survey.}

\item{peer_group}{Character. Peer group of the analysed portfolio.}
}
\value{
data.frame
}
\description{
Prepare data input for climate strategy metrics (initiatives) in the
scorecard section based on COP survey results. The input data sets are
pre-calculated and accessed via the directories pertaining to the given
user_id.
}
