% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_climate_strategy_scorecard.R
\name{prep_climate_strategy_scorecard_voting}
\alias{prep_climate_strategy_scorecard_voting}
\title{Prepare data input for climate strategy metrics (voting rights) in the scorecard}
\usage{
prep_climate_strategy_scorecard_voting(
  data,
  data_peers,
  peer_group = c("pensionfund", "assetmanager", "bank", "insurance", "other")
)
}
\arguments{
\item{data}{Data frame that contains pre-calculated climate strategy info on
use of voting rights on the individual user level, based on the
responses to the COP survey.}

\item{data_peers}{Data frame that contains pre-calculated climate strategy
info on peer level use of voting rights, based on the responses to
the COP survey.}

\item{peer_group}{Character. Peer group of the analysed portfolio.}
}
\value{
data.frame
}
\description{
Prepare data input for climate strategy metrics (voting rights) in the scorecard
section based on COP survey results. The input data sets are pre-calculated
and accessed via the directories pertaining to the given user_id.
}
