% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_executive_summary.R
\name{prep_data_executive_summary}
\alias{prep_data_executive_summary}
\title{Prepares data inputs needed for rendering the executive summary}
\usage{
prep_data_executive_summary(
  investor_name,
  portfolio_name,
  peer_group,
  start_year,
  scenario_source,
  scenario_selected,
  scenario_geography,
  equity_market,
  portfolio_allocation_method_equity,
  portfolio_allocation_method_bonds,
  green_techs,
  equity_results_portfolio,
  bonds_results_portfolio,
  peers_equity_results_aggregated,
  peers_bonds_results_aggregated,
  peers_equity_results_individual,
  peers_bonds_results_individual,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  audit_file,
  emissions_portfolio,
  index_dir
)
}
\arguments{
\item{investor_name}{Some description}

\item{portfolio_name}{Some description}

\item{peer_group}{Some description}

\item{start_year}{Some description}

\item{scenario_source}{Some description}

\item{scenario_selected}{Some description}

\item{scenario_geography}{Some description}

\item{equity_market}{Some description}

\item{portfolio_allocation_method_equity}{Some description}

\item{portfolio_allocation_method_bonds}{Some description}

\item{green_techs}{Some description}

\item{equity_results_portfolio}{Some description}

\item{bonds_results_portfolio}{Some description}

\item{peers_equity_results_aggregated}{Some description}

\item{peers_bonds_results_aggregated}{Some description}

\item{peers_equity_results_individual}{Some description}

\item{peers_bonds_results_individual}{Some description}

\item{indices_equity_results_portfolio}{Some description}

\item{indices_bonds_results_portfolio}{Some description}

\item{audit_file}{Some description}

\item{emissions_portfolio}{Some description}

\item{index_dir}{Directory containing \code{.rds} files with index benchmarks}
}
\value{
data.frame
}
\description{
Prepares data inputs needed for rendering the executive summary
}
