% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_diagram.R
\name{prep_diagram}
\alias{prep_diagram}
\title{Title}
\usage{
prep_diagram(audit_data = NULL, emissions_data = NULL, currency_exchange_value)
}
\arguments{
\item{audit_data}{Some arg}

\item{emissions_data}{Some arg}

\item{currency_exchange_value}{currency exchange rate (USD / currency)}
}
\value{
Some output
}
\description{
Title
}
