% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_emissions_scorecard.R
\name{prep_emissions_scorecard}
\alias{prep_emissions_scorecard}
\title{Title}
\usage{
prep_emissions_scorecard(
  emissions_data = NULL,
  audit_data,
  currency_exchange_value
)
}
\arguments{
\item{emissions_data}{Some arg}

\item{audit_data}{Some arg}

\item{currency_exchange_value}{Some arg}
}
\value{
Some output data
}
\description{
Title
}
