% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_exposures_scorecard.R
\name{prep_exposures_scorecard}
\alias{prep_exposures_scorecard}
\title{Prepare data input for plotting exposure chart on the climate score card}
\usage{
prep_exposures_scorecard(results_portfolio, scenario_selected = "NZE_2050")
}
\arguments{
\item{results_portfolio}{Data frame that contains pre-wrangled portfolio
level PACTA results from a PACTA for investors analysis.}

\item{scenario_selected}{Character. Must be a \code{scenario} featured in the
\code{scenario_thresholds} data set. Default is "1.5C-Unif", which is the 1.5°C
scenario from GECO2021.}
}
\value{
data.frame
}
\description{
Prepare data input for plotting exposure chart on the climate score card
based on PACTA for investors output files. These files must have been
wrangled with \code{prep_data_executive_summary()} before they can be passed to
this function.
}
