% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_exposures_survey.R
\name{prep_exposures_survey}
\alias{prep_exposures_survey}
\title{Prepare data input for plotting exposure to fossil fuels in survey section}
\usage{
prep_exposures_survey(
  results_portfolio,
  peers_results_aggregated,
  technology = c("coal", "oil", "gas"),
  asset_class = c("equity", "bonds")
)
}
\arguments{
\item{results_portfolio}{Data frame that contains pre-wrangled portfolio
level PACTA results from a PACTA for investors analysis.}

\item{peers_results_aggregated}{Data frame that contains pre-wrangled
aggregate peer group level PACTA results from a PACTA for investors
analysis.}

\item{technology}{Character. Must be of length 1 and either \code{coal} or
\code{oil} or \code{gas}.}

\item{asset_class}{Character. Must be of length 1 and either \code{equity} or
\code{bonds}.}
}
\value{
data.frame
}
\description{
Prepare data input for plotting exposure to fossil fuels in survey section
based on PACTA for investors output files. These files must have been
wrangled with \code{prep_data_executive_summary()} before they can be passed to
this function.
}
