% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fossil_bars.R
\name{prep_fossil_bars}
\alias{prep_fossil_bars}
\title{Prepare data input for plotting upstream fossil fuel exposure}
\usage{
prep_fossil_bars(
  results_portfolio,
  peers_results_aggregated,
  indices_results_portfolio,
  scenario_selected = "1.5C-Unif"
)
}
\arguments{
\item{results_portfolio}{Data frame that contains pre-wrangled portfolio
level PACTA results from a PACTA for investors analysis.}

\item{peers_results_aggregated}{Data frame that contains pre-wrangled
aggregate peer group level PACTA results from a PACTA for investors
analysis.}

\item{indices_results_portfolio}{Data frame that contains pre-wrangled
PACTA results for indices from a PACTA for investors analysis.}

\item{scenario_selected}{Character. Must be a \code{scenario} featured in the
\code{scenario_thresholds} data set. Defaults to \code{"1.5C-Unif"} from the GECO2021
scenario source.}
}
\value{
data.frame
}
\description{
Prepare data input for plotting upstream fossil fuel exposure based
on PACTA for investors output files. These files must have been wrangled with
\code{prep_data_executive_summary()} before they can be passed to this function.
}
