% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scatter.R
\name{prep_scatter}
\alias{prep_scatter}
\title{Title}
\usage{
prep_scatter(
  results_portfolio,
  peers_results_aggregated,
  peers_results_individual,
  indices_results_portfolio,
  scenario_source = "GECO2021",
  scenario_selected = "1.5C-Unif",
  asset_class = c("equity", "bonds")
)
}
\arguments{
\item{results_portfolio}{Some arg}

\item{peers_results_aggregated}{Some arg}

\item{peers_results_individual}{Some arg}

\item{indices_results_portfolio}{Some arg}

\item{scenario_source}{Some arg}

\item{scenario_selected}{Some arg}

\item{asset_class}{Some arg}
}
\value{
Some output
}
\description{
Title
}
