% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scores.R
\name{prep_scores}
\alias{prep_scores}
\title{Prepare data input for plotting aggregate climate scores}
\usage{
prep_scores(
  results_portfolio,
  peers_results_aggregated,
  asset_class = c("equity", "bonds"),
  scenario_source = "WEO2023"
)
}
\arguments{
\item{results_portfolio}{Data frame that contains pre-wrangled portfolio
level PACTA results from a PACTA for investors analysis.}

\item{peers_results_aggregated}{Data frame that contains pre-wrangled
aggregate peer group level PACTA results from a PACTA for investors
analysis.}

\item{asset_class}{Character. Must be either \code{equity} or
\code{bonds}.}

\item{scenario_source}{Character. Must be a
\code{scenario_source} featured in the \code{scenario_thresholds} data set.}
}
\value{
data.frame
}
\description{
Prepare data input for plotting aggregate climate scores based
on PACTA for investors output files. These files must have been wrangled with
\code{prep_data_executive_summary()} before they can be passed to this function.
}
