% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scores_scorecard.R
\name{prep_scores_exposure_scorecard}
\alias{prep_scores_exposure_scorecard}
\title{Prepare share of portfolio emissions covered by aggregate score analysis}
\usage{
prep_scores_exposure_scorecard(
  audit_data,
  currency_exchange_value,
  total_portfolio_value_curr,
  log_dir
)
}
\arguments{
\item{audit_data}{Data frame that contains pre-wrangled audit data}

\item{currency_exchange_value}{Numeric vector with exchange rate}

\item{total_portfolio_value_curr}{Numeric vector with total portfolio value
in target currency}

\item{log_dir}{Path to log file}
}
\value{
numeric
}
\description{
Prepare share of portfolio emissions covered by aggregate score analysis
}
