% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scores_scorecard.R
\name{prep_scores_scorecard}
\alias{prep_scores_scorecard}
\title{Prepare data input for plotting aggregate climate scores in the scorecard
section}
\usage{
prep_scores_scorecard(
  results_portfolio,
  asset_class = c("equity", "bonds"),
  scenario_source = "GECO2021"
)
}
\arguments{
\item{results_portfolio}{Data frame that contains pre-wrangled portfolio
level PACTA results from a PACTA for investors analysis.}

\item{asset_class}{Character. Must be either \code{equity} or
\code{bonds}.}

\item{scenario_source}{Character. Must be a
\code{scenario_source} featured in the \code{scenario_thresholds} data set.}
}
\value{
data.frame
}
\description{
Prepare data input for plotting aggregate climate scores in the scorecard
section based on PACTA for investors output files. These files must have been
wrangled with \code{prep_data_executive_summary()} before they can be passed to
this function. This simply calls \code{prep_scores()} and filters the appropriate
values.
}
