% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-re-exports.R
\name{rasterGrob}
\alias{rasterGrob}
\title{Render a raster object}
\usage{
rasterGrob(image)
}
\arguments{
\item{image}{Any R object that can be coerced to a raster object.}
}
\value{
A rastergrob grob.
}
\description{
See \code{grid::\link[grid:rasterGrob]{rasterGrob}} for details.
}
\keyword{internal}
