% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remaining_carbon_budgets}
\alias{remaining_carbon_budgets}
\title{Sector level carbon budgets for scenarios}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 5 columns.
}
\usage{
remaining_carbon_budgets
}
\description{
A data set containing the carbon budgets by sector until 2030.
}
\keyword{datasets}
