% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_executive_summary.R
\name{render_executive_summary}
\alias{render_executive_summary}
\title{Renders executive summary}
\usage{
render_executive_summary(
  data,
  language,
  output_dir,
  exec_summary_dir,
  survey_dir,
  real_estate_dir,
  file_name = "template.Rmd",
  investor_name,
  portfolio_name,
  peer_group,
  total_portfolio,
  scenario_selected,
  currency_exchange_value
)
}
\value{
a pdf document written to output_dir
}
\description{
Renders executive summary
}
