% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_thresholds}
\alias{scenario_thresholds}
\title{Scenario names mapped to temperature thresholds}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
scenario_thresholds
}
\description{
A data set containing the names of scenarios mapped to thresholds of likely
climate outcomes.
}
\keyword{datasets}
