% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_real_estate.R
\docType{data}
\name{scores_real_estate}
\alias{scores_real_estate}
\title{Help dataset for plotting real estate scores}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 5 columns.
}
\usage{
scores_real_estate
}
\description{
Dataset containing numeric values corresponding to alphabetic scores for
easier plotting.
}
\examples{
scores_real_estate
}
\keyword{internal}
