% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_diagram.R
\docType{data}
\name{toy_data_diagram}
\alias{toy_data_diagram}
\title{An example output data of \code{prep_diagram()}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 8 columns.
}
\usage{
toy_data_diagram
}
\description{
Dataset imitating the output of \code{prep_diagram()} function. Allows the
user to create plots without PACTA_analysis-like input data.
}
\examples{
toy_data_diagram
}
\seealso{
Other datasets: 
\code{\link{toy_data_alignment_table}},
\code{\link{toy_data_emissions_scorecard}},
\code{\link{toy_data_exposures_scorecard}},
\code{\link{toy_data_exposures_survey}},
\code{\link{toy_data_fossil_bars}},
\code{\link{toy_data_green_brown_bars}},
\code{\link{toy_data_scatter}},
\code{\link{toy_data_scores}}
}
\concept{datasets}
\keyword{datasets}
