% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_emissions_scorecard.R
\docType{data}
\name{toy_data_emissions_scorecard}
\alias{toy_data_emissions_scorecard}
\title{An example output data of \code{prep_emissions_scorecard()}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
toy_data_emissions_scorecard
}
\description{
Dataset imitating the output of \code{prep_emissions_scorecard()} function. Allows the
user to create plots without PACTA_analysis-like input data.
}
\examples{
toy_data_emissions_scorecard
}
\seealso{
Other datasets: 
\code{\link{toy_data_alignment_table}},
\code{\link{toy_data_diagram}},
\code{\link{toy_data_exposures_scorecard}},
\code{\link{toy_data_fossil_bars}},
\code{\link{toy_data_green_brown_bars}},
\code{\link{toy_data_scatter}},
\code{\link{toy_data_scores}}
}
\concept{datasets}
\keyword{datasets}
