% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_scatter.R
\docType{data}
\name{toy_data_scatter}
\alias{toy_data_scatter}
\title{An example output data of \code{prep_scatter()}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 6 columns.
}
\usage{
toy_data_scatter
}
\description{
Dataset imitating the output of \code{prep_scatter()} function. Allows the
user to create plots without PACTA_analysis-like input data.
}
\examples{
toy_data_scatter
}
\seealso{
Other datasets: 
\code{\link{toy_data_alignment_table}},
\code{\link{toy_data_diagram}},
\code{\link{toy_data_emissions_scorecard}},
\code{\link{toy_data_exposures_scorecard}},
\code{\link{toy_data_exposures_survey}},
\code{\link{toy_data_fossil_bars}},
\code{\link{toy_data_green_brown_bars}},
\code{\link{toy_data_scores}},
\code{\link{toy_data_scores_scorecard_re}}
}
\concept{datasets}
\keyword{datasets}
