% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_scores.R
\docType{data}
\name{toy_data_scores}
\alias{toy_data_scores}
\title{An example output data of \code{prep_scores()}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 5 columns.
}
\usage{
toy_data_scores
}
\description{
Dataset imitating the output of \code{prep_scores()} function. Allows the
user to create plots without PACTA_analysis-like input data.
}
\examples{
toy_data_scores
}
\seealso{
Other datasets: 
\code{\link{toy_data_alignment_table}},
\code{\link{toy_data_diagram}},
\code{\link{toy_data_emissions_scorecard}},
\code{\link{toy_data_exposures_scorecard}},
\code{\link{toy_data_exposures_survey}},
\code{\link{toy_data_fossil_bars}},
\code{\link{toy_data_green_brown_bars}},
\code{\link{toy_data_scatter}}
}
\concept{datasets}
\keyword{datasets}
