% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_scores_scorecard_re.R
\docType{data}
\name{toy_data_scores_scorecard_re}
\alias{toy_data_scores_scorecard_re}
\title{An example input data to \code{plot_scores_scorecard_real_estate()}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
toy_data_scores_scorecard_re
}
\description{
Dataset imitating the input to \code{plot_scores_scorecard_real_estate()} function.
Allows the user to create plots without PACTA_analysis-like input data.
}
\examples{
toy_data_scores_scorecard_re
}
\seealso{
Other datasets: 
\code{\link{toy_data_alignment_table}},
\code{\link{toy_data_diagram}},
\code{\link{toy_data_emissions_scorecard}},
\code{\link{toy_data_exposures_scorecard}},
\code{\link{toy_data_exposures_survey}},
\code{\link{toy_data_fossil_bars}},
\code{\link{toy_data_green_brown_bars}},
\code{\link{toy_data_scatter}},
\code{\link{toy_data_scores}}
}
\concept{datasets}
\keyword{datasets}
