% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-re-exports.R
\name{grid.arrange}
\alias{grid.arrange}
\title{Arrange multiple grobs on a page}
\usage{
grid.arrange(..., ncol)
}
\arguments{
\item{...}{grobs, gtables, ggplot or trellis objects.}

\item{ncol}{argument of gtable.}
}
\value{
draw on the current device.
}
\description{
See \code{gridExtra::\link[gridExtra:grid.arrange]{grid.arrange}} for details.
}
\keyword{internal}
