% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_net_zero_commitments.R
\name{prep_net_zero_commitments}
\alias{prep_net_zero_commitments}
\title{Prepare data input for share of net zero committed companies in scorecard}
\usage{
prep_net_zero_commitments(
  total_portfolio,
  peer_group = c("pensionfund", "assetmanager", "bank", "insurance", "other"),
  fin_data_net_zero_targets,
  peers_net_zero_commitment
)
}
\arguments{
\item{total_portfolio}{Data frame. Contains processed input of the PACTA for
Investors calculation.}

\item{peer_group}{Character. Peer group of the analysed portfolio.}

\item{fin_data_net_zero_targets}{Data frame. Contains information on which ISINs
belong to companies that have committed to SBTI net zero targets.}

\item{peers_net_zero_commitment}{DataFrame. Contains informations on results at peer group level to be compared with portfolio}
}
\value{
data.frame
}
\description{
Prepare data input for share of net zero committed companies in scorecard
based on PACTA for investors output files.
}
