% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scores_scorecard.R
\name{prep_scores_emissions_scorecard}
\alias{prep_scores_emissions_scorecard}
\title{Prepare share of portfolio emissions covered by aggregate score analysis}
\usage{
prep_scores_emissions_scorecard(emissions_data, log_dir)
}
\arguments{
\item{emissions_data}{Data frame that contains pre-wrangled emissions data}

\item{log_dir}{Path to log file}
}
\value{
numeric
}
\description{
Prepare share of portfolio emissions covered by aggregate score analysis
}
